refactor: likeMutation 분리
